Added more TODO items
The ProteusCreatureTest project was added. I does a few simple tests of the system and they are working at this point. The parts tested so far are: Adding/replacing bodyparts, abilities, and (passive) effects. A simple algorithm for max health has been made, (strength * .1 + endurance * .75) * 100. This makes for a large amount of health. Will need work. Want to make a table detailing the possible amounts of health. I'll either do it in code or in a google doc file.

Many more tests to be done and more simplifying to do. Adding more constructors as I see fit.

Ryan: start taking a crack at the TODO items, watch out that some of them might have been done already.

In bodypart ClassType has been split into ClassType, PartType, ClassPartType, where ClassPartType is not what Class Type was and  Class/PartType enums are not individual elements.

More to come